Enable the location option for copy rule.
The location option can be used in order to set explicitly a
directory or a file to copy. Note that a file is mostly false
because it will put this file in the root directory of the
package.
If this option is omitted, then the current location available
in the URI option will be used. For example, the location is
mostly useless with self: protocol.