Add initial complete support for make dependencies
The packages in the make dependency sction, are 'make', 'build' and
'install' before that the current package will be 'make'.

It's not optimal, because some steps are a bit slow. We should skip the
build step, if it's already built. Same with the install step. If it's
not rebuilt, we should check if it's already installed.