Rework external setting by using a distribution name
Instead of true/false, it's a distribution name because the external
package is referenced in the sources.list of a specific distribution.
This package must be installed in this distribution.

The mail problem is that we must use a build dependency to specify
this sort of external dependency because the install dependencies
must be only sources.