Call bumpIfNecessary in all cases
This commit fixes the case where a package with this version is already
available in the archive reposiroty, but it's not the case for the
working repository. Then, a bump must be performed otherwise two
different packages with the same version exist at the same time.

Note that this commit add support for the distribution because to have
different versions when using specific distribution definitions (with
+ symbol).